+++ tests/single_card_tests/transformer/test_hybrid_mla_doc_equivalence.py
@@ -18,10 +18,10 @@ Validation agent A4. The gold standard proven here: N documents PACKED into a
 single 8192-style sequence must equal running every document ALONE (same
 weights, same content), elementwise on outputs AND gradients. The hybrid MLA
 (``csa_compress_ratios == -2``) layers run one of two attentions, selected by
-``un_absorbed_mqa`` and, within it, by whether the sublayers spec carries an
+``non_absorbed_mqa`` and, within it, by whether the sublayers spec carries an
 indexer:
 
-* dense MHA (``un_absorbed_mqa=False``) -- the fp32 ``_dense_reference`` below.
+* dense MHA (``non_absorbed_mqa=False``) -- the fp32 ``_dense_reference`` below.
 * the indexer-less :class:`MQALatentAttention` dense path -- absorbed MQA,
   mathematically equal to MHA (spec ``indexer=None``, a test-only construction
   since production always builds the indexer).
@@ -30,9 +30,10 @@ indexer:
 Each is checked with the model-wide learnable per-head attention sink both ON
 and OFF.
 
-The file is intentionally self-contained (it re-implements the small helpers of
-``test_mqa_latent_attention.py`` instead of importing that test module) so the
-proof does not depend on another test's internals.
+The shared direct-construction fixtures (geometry, stub sublayers, config
+factory, module builder, fp32 dense reference) live in ``hybrid_mla_utils.py``,
+a plain helper module rather than a test module, so the proof still does not
+depend on another test's internals.
 
 GPU-gated classes require the SM100+ FlashMLA sparse fwd + cuDNN DSA kernels;
 the pure metadata / boundary-attack classes run on CPU.
@@ -42,8 +43,6 @@ import unittest
 
 import numpy as np
 import paddle
-import paddle.nn.functional as F
-from paddle.distributed.fleet.meta_parallel import LayerSpec
 
 from paddlefleet.transformer.csa_attention import (
     _build_valid_range_from_doc_bounds,
@@ -51,209 +50,27 @@ from paddlefleet.transformer.csa_attention import (
     _derive_csa_doc_boundaries,
     _validate_csa_docmask_shape,
 )
-from paddlefleet.transformer.dsa_attention import (
-    DSAIndexer,
-    DSAIndexerLossLoggingHelper,
-    DSAIndexerSublayersSpec,
+from paddlefleet.transformer.dsa_attention import DSAIndexerLossLoggingHelper
+
+from .hybrid_mla_utils import (
+    _CAPTURED,
+    _GPU,
+    DV,
+    HIDDEN,
+    INDEX_TOPK,
+    K_CHANNELS,
+    Q_LORA,
+    V_HEAD_DIM,
+    WINDOW,
+    H,
+    _build_module,
+    _check_index_invariants,
+    _create_mqa_config,
+    _dense_reference,
+    _doc_meta,
+    _make_inputs,
+    _row_end,
 )
-from paddlefleet.transformer.enums import AttnMaskType
-from paddlefleet.transformer.mqa_latent_attention import (
-    MQALatentAttention,
-    MQALatentAttentionSublayersSpec,
-)
-from paddlefleet.transformer.transformer_config import TransformerConfig
-from paddlefleet.utils import init_method_normal, scaled_init_method_normal
-
-# ---------------------------------------------------------------------------
-# Geometry (identical hard requirements to test_mqa_latent_attention.py).
-# ---------------------------------------------------------------------------
-H = 8
-DK = 576
-DV = 512
-V_HEAD_DIM = 64
-K_CHANNELS = 256
-WINDOW = 128
-INDEX_TOPK = 128
-INDEX_HEADS = 64
-INDEX_HEAD_DIM = 128
-HIDDEN = 256
-Q_LORA = 128
-
-
-class BiasedLinear(paddle.nn.Layer):
-    def __init__(self, in_features, out_features, **kwargs):
-        super().__init__()
-        self.linear = paddle.nn.Linear(in_features, out_features)
-
-    def forward(self, x):
-        if x.dtype != self.linear.weight.dtype:
-            x = x.cast(self.linear.weight.dtype)
-        return self.linear(x), self.linear.bias
-
-
-class LayerNormStub(paddle.nn.Layer):
-    def __init__(
-        self,
-        hidden_size=None,
-        eps=None,
-        normalized_shape=None,
-        epsilon=None,
-        **kwargs,
-    ):
-        super().__init__()
-        size = hidden_size if hidden_size is not None else normalized_shape
-        self.eps = (
-            eps
-            if eps is not None
-            else (epsilon if epsilon is not None else 1e-5)
-        )
-        self.weight = paddle.nn.Parameter(paddle.ones([size]))
-        self.bias = paddle.nn.Parameter(paddle.zeros([size]))
-
-    def forward(self, x):
-        mean = x.mean(axis=-1, keepdim=True)
-        var = x.var(axis=-1, keepdim=True, unbiased=False)
-        x = (x - mean) / paddle.sqrt(var + self.eps)
-        return x * self.weight + self.bias
-
-
-def _create_mqa_config(mode="mqa", loss_coeff=0.0, num_hidden_layers=2):
-    """dsv4_hybrid config for a ``csa_compress_ratios == -2`` layer.
-
-    ``mode`` is a test-only convenience: both ``"mqa"`` (dense, indexer-less)
-    and ``"mqa_dsa"`` (DSA) set ``un_absorbed_mqa=True``; the dense/sparse
-    distinction is expressed by whether ``_build_module`` attaches an indexer to
-    the sublayers spec, mirroring the production source.
-    """
-    config = TransformerConfig(
-        num_hidden_layers=num_hidden_layers,
-        hidden_size=HIDDEN,
-        num_attention_heads=H,
-    )
-    config.num_key_value_heads = H
-    config.head_dim = K_CHANNELS
-    config.experimental_attention_variant = "dsv4_hybrid"
-    config.un_absorbed_mqa = True
-    # Test-only marker read by ``_build_module``: production always builds the
-    # indexer when ``un_absorbed_mqa`` is set, so the indexer-less dense path is
-    # reachable only by constructing the layer directly with ``indexer=None``.
-    config._build_dsa_indexer = mode == "mqa_dsa"
-    config.hybrid_mla_q_lora_rank = Q_LORA
-    config.hybrid_mla_kv_lora_rank = DV
-    config.hybrid_mla_qk_nope_head_dim = 192
-    config.hybrid_mla_qk_rope_head_dim = 64
-    config.hybrid_mla_v_head_dim = V_HEAD_DIM
-    config.hybrid_mla_num_attention_heads = H
-    config.hybrid_mla_num_key_value_heads = H
-    # Model-wide indexer dims (HF json aliases index_n_heads / index_head_dim /
-    # index_topk), shared with the CSA layers.
-    config.dsa_index_n_heads = INDEX_HEADS
-    config.dsa_index_head_dim = INDEX_HEAD_DIM
-    config.dsa_index_topk = INDEX_TOPK
-    config.csa_window_size = WINDOW
-    config.dsa_indexer_loss_coeff = loss_coeff
-    config.dsa_indexer_use_sparse_loss = True
-    config.dsa_indexer_rotary_interleaved = False
-    config.rope_type = "rope"
-    config.rope_theta = 10000.0
-    config.rotary_interleaved = False
-    config.rotary_percent = 1.0
-    config.apply_rope_fusion = False
-    config.num_nextn_predict_layers = 0
-    config.mtp_num_layers = 0
-    config.init_method = init_method_normal(0.02)
-    config.output_layer_init_method = scaled_init_method_normal(0.02, 1, 2.0)
-    config.rms_norm_eps = 1e-5
-    config.context_parallel_size = 1
-    config.sequence_parallel = False
-    return config
-
-
-_CAPTURED = []
-
-
-class RecordingMQA(MQALatentAttention):
-    """Captures the ``token_indices`` handed to the sparse kernel."""
-
-    def _sparse_attn(self, query, kv, token_indices, sm_scale, d_v):
-        _CAPTURED.append(token_indices.numpy().copy())
-        return super()._sparse_attn(query, kv, token_indices, sm_scale, d_v)
-
-
-def _build_module(config, layer_number=1, bf16=False, sink=None, is_mtp=False):
-    indexer = None
-    if getattr(config, "_build_dsa_indexer", False):
-        indexer = LayerSpec(
-            layer=DSAIndexer,
-            sublayers_spec=DSAIndexerSublayersSpec(
-                linear_wq_b=BiasedLinear,
-                linear_wk=BiasedLinear,
-                k_norm=LayerNormStub,
-                linear_weights_proj=BiasedLinear,
-            ),
-            extra_kwargs={"is_hybrid_mla_indexer": True},
-        )
-    module = RecordingMQA(
-        config=config,
-        sublayers_spec=MQALatentAttentionSublayersSpec(indexer=indexer),
-        layer_number=layer_number,
-        attn_mask_type=AttnMaskType.causal,
-        attention_type="self",
-        k_channels=K_CHANNELS,
-        is_mtp_layer=is_mtp,
-    )
-    if bf16:
-        module.to(dtype="bfloat16")
-    if sink is not None:
-        module.softmax_offset = module.create_parameter(
-            shape=[H],
-            dtype="bfloat16" if bf16 else "float32",
-            default_initializer=paddle.nn.initializer.Assign(
-                np.asarray(sink, dtype="float32")
-            ),
-        )
-    return module
-
-
-def _dsa_kernels_available():
-    if not paddle.is_compiled_with_cuda():
-        return False
-    try:
-        from paddlefleet.cudnn_ops.block_sparse_mqa_dsa import is_dsa_available
-
-        return bool(is_dsa_available())
-    except Exception:
-        return False
-
-
-_GPU = unittest.skipUnless(
-    _dsa_kernels_available(),
-    "requires SM100+ FlashMLA sparse fwd + cuDNN DSA bwd kernels",
-)
-
-
-def _row_end(doc_lens, seqlen):
-    """``[1, 1, s, 1]`` int32 exclusive per-token document end row.
-
-    The trailing gap (``sum(doc_lens) < seqlen``) becomes one final document
-    ending at ``seqlen``, so every row is valid.
-    """
-    out = np.empty([seqlen], dtype="int32")
-    pos = 0
-    for length in doc_lens:
-        end = pos + length
-        out[pos : min(end, seqlen)] = end
-        pos = end
-        if pos >= seqlen:
-            break
-    if pos < seqlen:
-        out[pos:] = seqlen
-    return paddle.to_tensor(out).reshape([1, 1, seqlen, 1])
-
-
-def _doc_meta(row_end, seqlen):
-    doc_start, _, is_valid, _, _ = _derive_csa_doc_boundaries(row_end, seqlen)
-    return doc_start.numpy(), is_valid.numpy()
 
 
 def _doc_segments(layout, seqlen):
@@ -269,20 +86,6 @@ def _doc_segments(layout, seqlen):
     return list(zip(starts, lens))
 
 
-def _make_inputs(seqlen, seed=0):
-    paddle.seed(seed)
-    query = (paddle.randn([1, seqlen, H, DK]) * 0.5).cast("bfloat16")
-    key = (paddle.randn([1, seqlen, 1, DK]) * 0.5).cast("bfloat16")
-    w_v = (paddle.randn([DV, H, V_HEAD_DIM]) * 0.05).cast("bfloat16")
-    return query, key, w_v
-
-
-def _rel(actual, expected):
-    a = actual.cast("float32")
-    e = expected.cast("float32")
-    return float((a - e).norm() / e.norm().clip(min=1e-12))
-
-
 def _relerr_maxmean(actual, expected, eps=1e-6):
     """Elementwise relative error stats: ``(norm_rel, max_abs, max_rel)``."""
     a = np.asarray(actual, dtype=np.float64)
@@ -295,79 +98,6 @@ def _relerr_maxmean(actual, expected, eps=1e-6):
         float((absdiff / (np.abs(e) + eps)).max()) if absdiff.size else 0.0
     )
     return norm_rel, max_abs, max_rel
-
-
-def _dense_reference(query, key, w_v, row_end, scale, sink=None):
-    """Per-document full-causal attention on the latent, computed in fp32.
-
-    This *is* the mathematical ``mha`` path (dense MLA on the absorbed latent):
-    absorption is exactly score preserving, so ``mha``/``mqa``/``mqa_dsa`` must
-    all reduce to this reference. ``sink`` is a ``[H]`` per-head logit appended
-    as one extra value-less softmax column (drains probability mass only).
-    """
-    seqlen = int(query.shape[1])
-    doc_start, is_valid = _doc_meta(row_end, seqlen)
-    pos = np.arange(seqlen)
-    allowed = (
-        (pos[None, :] <= pos[:, None])
-        & (pos[None, :] >= doc_start[:, None])
-        & is_valid[:, None]
-    )
-    q = query[0].cast("float32")
-    k = key.squeeze(2)[0].cast("float32")
-    scores = paddle.einsum("shd,td->sht", q, k) * scale
-    keep = paddle.to_tensor(allowed).unsqueeze(1)
-    scores = paddle.where(keep, scores, paddle.full_like(scores, -1e30))
-    if sink is None:
-        probs = F.softmax(scores, axis=-1)
-    else:
-        sink_col = paddle.to_tensor(np.asarray(sink, dtype="float32")).reshape(
-            [1, H, 1]
-        )
-        sink_col = paddle.expand(sink_col, [seqlen, H, 1])
-        probs = F.softmax(paddle.concat([scores, sink_col], axis=-1), axis=-1)
-        probs = probs[:, :, :seqlen]
-    ctx = paddle.einsum("sht,tl->shl", probs, k[:, :DV])
-    out = paddle.einsum("shl,lhv->shv", ctx, w_v.cast("float32"))
-    row_ok = paddle.to_tensor(is_valid).cast("float32").reshape([seqlen, 1, 1])
-    return (out * row_ok).reshape([1, seqlen, H * V_HEAD_DIM])
-
-
-def _check_index_invariants(test, indices, row_end, seqlen, expect_full=False):
-    """Assert the per-row selected column set is sound (task 2/3).
-
-    Invariants: no duplicate, every column causal + in-document, the forced
-    ``WINDOW`` columns always present and clipped at ``doc_start``, no
-    out-of-range column, pad rows select nothing.
-    """
-    doc_start, is_valid = _doc_meta(row_end, seqlen)
-    for q in range(seqlen):
-        cols = indices[0, q]
-        cols = cols[cols >= 0].tolist()
-        test.assertEqual(
-            len(cols), len(set(cols)), f"row {q}: duplicate column"
-        )
-        test.assertTrue(
-            all(0 <= c < seqlen for c in cols), f"row {q}: out-of-range column"
-        )
-        if not is_valid[q]:
-            test.assertEqual(cols, [], f"pad row {q} must select nothing")
-            continue
-        start = int(doc_start[q])
-        test.assertTrue(
-            all(start <= c <= q for c in cols),
-            f"row {q}: non-causal or cross-document column",
-        )
-        window = set(range(max(start, q - WINDOW + 1), q + 1))
-        test.assertEqual(
-            window - set(cols), set(), f"row {q}: lost forced-window columns"
-        )
-        if expect_full:
-            test.assertEqual(
-                set(cols),
-                set(range(start, q + 1)),
-                f"row {q}: not the full causal set",
-            )
 
 
 # Adversarial layouts required by the task, as ``(layout, seqlen)`` pairs.
